updated two build.js to use path module

+++ build.js
@@ -1,23 +1,20 @@
 const fs = require('fs');
+const path = require('path');
 const uglify = require('uglify-js');
 
-const sourceDirectory = 'src/';
-const destinationDirectory = 'build/';
+const sourceDirectory = 'src';
+const destinationDirectory = 'build';
 
-try {
+if (path.existsSync(destinationDirectory)) {
     var oldFileNames = fs.readdirSync(destinationDirectory);
     for (var i = 0; i < oldFileNames.length; i++) {
-        var oldFileName = destinationDirectory + oldFileNames[i];
+        var oldFileName = path.join(destinationDirectory, oldFileNames[i]);
         fs.unlinkSync(oldFileName);
     }
     fs.rmdirSync(destinationDirectory);
     console.log('removed build directory');
-} catch (e) {
-    if (e.code == 'ENOENT') {
-        console.log('build directory doesn\'t exist');
-    } else {
-        throw e;
-    }
+} else {
+    console.log('build directory doesn\'t exist');
 }
 
 fs.mkdirSync(destinationDirectory, '755');
@@ -27,9 +24,9 @@ var fileNames = fs.readdirSync(sourceDirectory);
 var allFiles = [];
 
 for (var i = 0; i < fileNames.length; i++) {
-    var sourceFileName = sourceDirectory + fileNames[i];
-    var destinationFileName = destinationDirectory + fileNames[i];
-    var destinationCompressedFileName = destinationDirectory + fileNames[i].replace(/.js$/, '-min.js');
+    var sourceFileName = path.join(sourceDirectory, fileNames[i]);
+    var destinationFileName = path.join(destinationDirectory, fileNames[i]);
+    var destinationCompressedFileName = path.join(destinationDirectory, fileNames[i].replace(/.js$/, '-min.js'));
     
     var sourceFile = String(fs.readFileSync(sourceFileName));
     allFiles.push(sourceFile);
@@ -44,7 +41,7 @@ for (var i = 0; i < fileNames.length; i++) {
 
 var composedFile = allFiles.join('');
 var compressedComposedFile = uglify(composedFile);
-fs.writeFileSync(destinationDirectory + 'jshelpers.js', composedFile);
-console.log('composed all files into ' + destinationDirectory + 'jshelpers.js');
-fs.writeFileSync(destinationDirectory + 'jshelpers-min.js', compressedComposedFile);
-console.log('compressed all files into ' + destinationDirectory + 'jshelpers-min.js');
+fs.writeFileSync(path.join(destinationDirectory, 'jshelpers.js'), composedFile);
+console.log('composed all files into ' + path.join(destinationDirectory, 'jshelpers.js'));
+fs.writeFileSync(path.join(destinationDirectory, 'jshelpers-min.js'), compressedComposedFile);
+console.log('compressed all files into ' + path.join(destinationDirectory, 'jshelpers-min.js'));
